+++ src/number/isInteger.js
@@ -1,0 +1,8 @@
+'use strict';
+
+import isNumber from './is';
+
+export default function (val) {
+    if (!isNumber(val)) return false;
+    return Number.isInteger(val);
+}